requirements.txt
Django==2.2.11
django-braces==1.13.0
django-suit==0.2.28
django-easy-select2==1.5.2
social-auth-app-django==3.1.0
martor==1.4.1
pillow==7.1.0
django-import-export==1.2.0
djongo==1.3.1